refactor: split components of the game
